Integrate previous tools
