+++ openfisca_us/reforms/__init__.py
@@ -5,5 +5,8 @@ A reform is a set of modifications to be applied to a reference tax and benefit 
 
 See https://openfisca.org/doc/key-concepts/reforms.html
 """
-from openfisca_us.reforms.parametric import parametric_reform, reform_from_file
-from openfisca_us.reforms.tools import parametric, structural
+from openfisca_us.reforms.tools.parametric import (
+    parametric_reform,
+    reform_from_file,
+    set_parameter,
+)
