+++ openfisca_us/reforms/tools/parametric.py
@@ -72,3 +72,19 @@ def change_parameter(
             self.modify_parameters(modifier)
 
     return reform
+
+
+def parametric_reform(modifier_func):
+    class reform(Reform):
+        def apply(self):
+            self.modify_parameters(modifier_func)
+
+    return reform
+
+
+def reform_from_file(filepath: str):
+    def replace_parameters(parameters):
+        parameters = load_parameter_file(filepath)
+        return parameters
+
+    return parametric_reform(replace_parameters)
